examples/mnist_app/pyproject.toml
[project]
name = "mnist-app"
version = "0.1.0"
description = "MNIST digit recognition web app using dpl"
requires-python = ">=3.13"
dependencies = [
    "dpl",
    "streamlit>=1.30.0",
    "streamlit-drawable-canvas>=0.9.3",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[tool.uv.sources]
dpl = { path = "../..", editable = true }
